did some cleanup.  removed content from blog area that was being used for testing
