+++ client/src/components/About/About.css
@@ -7,7 +7,3 @@
 ul#feature-list {
   list-style-type: none;
 }
-
-/* #feature-list li i {
-  color: green;
-} */
