+++ client/src/components/Articles/Articles.js
@@ -1,37 +1,39 @@
-// import React from "react";
 import React, { Component } from 'react';
+import { Consumer } from '../../context';
+
 // import Wrapper from "../Wrapper";
 // import CardBtn from "../CardBtn";
+import ArticleItem from "../ArticleItem";
 import "./Articles.css";
 import moment from 'moment';
 
 // load up a series of button definitions
-const cardButtons = [
-  {
-    "id": 1,
-    "datavalue": "read",
-    "title": "Read",
-    "icon": "fas fa-book-open"
-  },
-  {
-    "id": 2,
-    "datavalue": "save",
-    "title": "Save",
-    "icon": "fas fa-thumbtack"
-  },
-  {
-    "id": 3,
-    "datavalue": "delete",
-    "title": "Delete",
-    "icon": "far fa-trash-alt"
-  },
-  {
-    "id": 4,
-    "datavalue": "comment",
-    "title": "Comment",
-    "icon": "far fa-comments fa-lg"
-  }
-];
+// const cardButtons = [
+//   {
+//     "id": 1,
+//     "datavalue": "read",
+//     "title": "Read",
+//     "icon": "fas fa-book-open"
+//   },
+//   {
+//     "id": 2,
+//     "datavalue": "save",
+//     "title": "Save",
+//     "icon": "fas fa-thumbtack"
+//   },
+//   {
+//     "id": 3,
+//     "datavalue": "delete",
+//     "title": "Delete",
+//     "icon": "far fa-trash-alt"
+//   },
+//   {
+//     "id": 4,
+//     "datavalue": "comment",
+//     "title": "Comment",
+//     "icon": "far fa-comments fa-lg"
+//   }
+// ];
 
 class Articles extends Component {
 
@@ -41,7 +43,7 @@ class Articles extends Component {
 
   // load the images from the json array
   state = {
-    cardButtons: cardButtons,
+    // cardButtons: cardButtons,
     comments: []
   };
   
@@ -53,6 +55,7 @@ class Articles extends Component {
     return (
 
       <div className="card card-body mb-3">
+
         <h6 className="text-left">{this.props.title} 
           <i className="fas fa-caret-down ml-2" style={{ cursor: 'pointer' }}></i>
           <i className="fas fa-times app-color-2" style={{cursor: 'pointer', float: 'right'}}></i>
@@ -60,7 +63,29 @@ class Articles extends Component {
             <i className="fas fa-pencil-alt app-color-4" style={{cursor: 'pointer', float: 'right', marginRight: '1rem'}}></i>
           </a>
         </h6>
+
         <small className="text-left text-muted">Written by {this.props.author} on {this.formatDate(this.props.date)}</small>
+
+        <Consumer>
+          { (value) => (    
+            <React.Fragment>
+              {value.articles.map(articleItem => (
+                <ArticleItem
+                  id={articleItem._id}
+                  key={articleItem._id}
+                  title={articleItem.title}
+                  date={articleItem.date}
+                  url={articleItem.url}
+                  summary={articleItem.summary}
+                  author={articleItem.author}
+                  comments={articleItem.comments}
+                />
+
+              ))}
+            </React.Fragment>
+          ) }
+        </Consumer>
+
       </div>
 
     )
