--- client/src/components/Card/Card.css
@@ -1,28 +1,0 @@
-/* .card {
-  background-size: cover;
-  background-position: center;
-  background-clip: content-box;
-  border: 6px solid white;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
-  position: relative;
-  margin: 0 auto;
-  margin-top: 20px;
-} */
-
-.card.sized-card {
-  background-size: cover;
-  background-position: center;
-  background-clip: content-box;
-  border: 6px solid white;
-  box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
-  position: relative;
-  margin: 0 auto;
-  margin-top: 20px;
-  height: 500px;
-  width: 500px;
-}
-
-/* remove the blue outline around buttons */
-.card.sized-card button:focus {
-  outline:0;
-}
