--- client/src/components/CardBtn/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CardBtn";
